Rename features to controllers.
