+++ app.js
@@ -6,7 +6,7 @@ const cookieParser = require('cookie-parser');
 const session = require('express-session');
 const RedisStore = require('connect-redis')(session);
 
-const initAuthMiddleware = require('./features/login/init-auth-middleware');
+const initAuthMiddleware = require('./controllers/login/init-auth-middleware');
 const indexRouter = require('./routes/index');
 
 const redisStoreConfig = {
