+++ controllers/logout/commands/logout.js
