+++ controllers/register/commands/load-page.js
