+++ controllers/register/commands/verify-request-body.js
