+++ controllers/reset-password/routes.js
